Update package references and refactor code for improved functionality and consistency
- Updated SQLitePCLRaw.bundle_e_sqlite3 from version 3.0.3 to 3.0.5.
- Upgraded Microsoft.AspNetCore packages from version 10.0.11 to 10.0.12 across multiple projects.
- Refactored AccountRegisterController to remove password from ViewData for security.
- Enhanced PaymentController to include CustomerEmail in payment requests.
- Changed ProductController to use long for categoryId instead of int for better compatibility.
- Updated ProductSearchRequest and related DTOs to use long instead of int for IDs.
- Modified specifications to handle long IDs and added null checks for robustness.
- Improved OrderRepository and CustomerRepository to set customer email based on user data.
- Adjusted various views to bind ViewData correctly and removed unnecessary scripts.
- Updated logging package reference to version 10.0.12.
- Refined Razor views for better user experience and security.

+++ eCommerce.Storefront.Controllers/Controllers/ProductController.cs
@@ -20,7 +20,7 @@ namespace eCommerce.Storefront.Controllers.Controllers
     {
         private readonly IConfiguration _configuration = configuration;
 
-        public async Task<IActionResult> GetProductsByCategory(int categoryId)
+        public async Task<IActionResult> GetProductsByCategory(long categoryId)
         {
             var productSearchRequest = GenerateInitialProductSearchRequestFrom(categoryId);
             var response = await _cachedProductCatalogService.GetProductsByCategoryAsync(productSearchRequest);
@@ -47,7 +47,7 @@ namespace eCommerce.Storefront.Controllers.Controllers
             return productSearchResultView;
         }
 
-        private GetProductsByCategoryRequest GenerateInitialProductSearchRequestFrom(int categoryId)
+        private GetProductsByCategoryRequest GenerateInitialProductSearchRequestFrom(long categoryId)
         {
             var productSearchRequest = new GetProductsByCategoryRequest
             {
