+++ eCommerce.Storefront.Controllers/DTOs/ProductSearchRequest.cs
@@ -5,10 +5,10 @@ namespace eCommerce.Storefront.Controllers.DTOs
 {
     public class ProductSearchRequest
     {
-        public int CategoryId { get; set; }
-        public int[] ColorIds { get; set; }
-        public int[] SizeIds { get; set; }
-        public int[] BrandIds { get; set; }
+        public long CategoryId { get; set; }
+        public long[] ColorIds { get; set; }
+        public long[] SizeIds { get; set; }
+        public long[] BrandIds { get; set; }
         public ProductsSortBy SortBy { get; set; }
         public IEnumerable<RefinementGroup> RefinementGroups { get; set; }
         public int Index { get; set; }
