+++ eCommerce.Storefront.Controllers/DTOs/RefinementGroup.cs
@@ -3,6 +3,6 @@ namespace eCommerce.Storefront.Controllers.DTOs
     public class RefinementGroup
     {
         public int GroupId { get; set; }
-        public int[] SelectedRefinements { get; set; }
+        public long[] SelectedRefinements { get; set; }
     }
 }
