+++ eCommerce.Storefront.Controllers/ViewModels/ProductCatalog/ProductSearchResultView.cs
@@ -11,7 +11,7 @@ namespace eCommerce.Storefront.Controllers.ViewModels.ProductCatalog
         }
 
         public string SelectedCategoryName { get; set; }
-        public int SelectedCategory { get; set; }
+        public long SelectedCategory { get; set; }
         public IEnumerable<RefinementGroup> RefinementGroups { get; set; }
         public int NumberOfTitlesFound { get; set; }
         public int TotalNumberOfPages { get; set; }
